refactor: Create Hero component
