+++ src/components/Hero.scss
@@ -1,3 +1,7 @@
+.hero {
+  background-color: var(--bg-hero);
+}
+
 .huge-header {
   text-align: center;
 }
