+++ src/styles.scss
@@ -103,9 +103,6 @@ footer.navbar {
     color: inherit !important;
   }
 }
-.hero {
-  background-color: var(--bg-hero);
-}
 svg.fill-me-anyway {
   fill: var(--text);
 }
